feat: Improved the architecture pattern
Improved the architecture pattern by making it more idiomatic and coherent

Changes made:
- Go module renaming
- Renaming the "server" binary to "http"
- Updated documentation
- Package organization was restructured following the [Go Project Layout Standard](https://github.com/golang-standards/project-layout)
- Builds tags added to compile only what is necessary
