+++ README.md
@@ -7,7 +7,6 @@ The purpose of the project is to provide an API to manage recipes and ingredient
 [OpenAPI](./docs/openapi.yaml)
 
 ### Getting started
-
 The API is composed of two parts, the first is an `REST API` that helps to manage the searches and storage related to recipes and the second is a `CLI` helps to fill the recipe storage.
 
 ### How to use the CLI
@@ -25,7 +24,7 @@ make cli
 ###### Compile
 Follow the instructions below to compile the binary file for the `REST API` in the `/build` directory
 ```shell
-make server
+make http
 ```
 ###### Load environment variables
 ```shell
@@ -33,26 +32,45 @@ export $(grep -v ^# .env.example)
 ```
 ###### Start the server
 ```shell
-./build/server
+./build/http
 ```
 
 ### Scripts
 This project contains some bash scripts to help to make some operations like compile.
-
 [See bash scripts](./scripts)
 
-### Architecture pattern
-This project implements architecture pattern [ports and adapters](https://alistair.cockburn.us/hexagonal-architecture)
+### Architecture
+###### Pattern (also style)
+This project implements architecture pattern [Ports and Adapters](https://alistair.cockburn.us/hexagonal-architecture)
+###### Decisions
+**Vertical Slicing**
+
+Interpreting what [Vertical Slicing](https://en.wikipedia.org/wiki/Vertical_slice) says, I decided to make one package per feature and put a little of each layer in each package.
+
+**Go Project Layout Standard**
+
+I decided to follow the [Go Project Layout Standard](https://github.com/golang-standards/project-layout).
+
+**Compile only what is required**
+
+According to the theory of hexagonal architecture, it is possible to have *n* adapters for different external signals (http, gRPC, command line).
+
+*For example* one use case is to create a recipe, but the instruction comes from a http request or a kafka message.
+
+So I decided to compile a binary to handle each signal.
+
+###### Package tree
 ```
 .
 ├── cmd
-│   └── {binary}
-│       └── container                    (Dependency injection)
+│   └── {binary}
 ├── internal
-│   └── {feature}                        (Business logic and adpaters for a specific feature)
-│       ├── business                     (Use cases, business rules and entities)
-│       └── infrastructure
-│           ├── input                    (Everything related to "drive" adapters)
-│           └── output                   (Everything related to "driven" adapters)
-└── pkg                                  (Global and public code, potentially libraries)
+│   ├── app
+│   │   └── {feature}
+│   │       ├── business (Use cases, rules, data models and ports)
+│   │       └── infrastructure
+│   │           ├── input  (Everything related to "driving" adapters)
+│   │           └── output (Everything related to "driven" adapters)
+│   └── container (DI container)
+└── pkg (Public and global code, potencially libraries)
 ```
