--- cmd/server/server.go
@@ -1,56 +1,0 @@
-package main
-
-import (
-	"context"
-	"github.com/labstack/echo/v4"
-	"github.com/yael-castro/cb-search-engine-api/cmd/server/container"
-	"log"
-	"os"
-	"os/signal"
-	"time"
-)
-
-const defaultPort = "8080"
-
-func main() {
-	port := os.Getenv("PORT")
-	if port == "" {
-		port = defaultPort
-	}
-
-	// Sets logger flags
-	log.SetFlags(log.Flags() | log.Llongfile)
-
-	// Creates main context
-	ctx, cancelFunc := context.WithCancel(context.Background())
-
-	signalCh := make(chan os.Signal)
-	signal.Notify(signalCh, os.Interrupt)
-
-	go func() {
-		<-signalCh
-		cancelFunc()
-	}()
-
-	e := echo.New()
-
-	// DI container input action!
-	err := container.Inject(ctx, e)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	go func() {
-		<-ctx.Done()
-
-		ctx, cancelFunc := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancelFunc()
-
-		_ = e.Shutdown(ctx)
-	}()
-
-	// Runs the server
-	log.Printf("%v\n", e)
-	log.Printf("Server version '%s' is running on port '%s'\n", container.GitCommit, port)
-	log.Fatal(e.Start(":" + port))
-}
